+++ Automation/ScriptingEngine/ScriptableComponents/Components/Configurator.cs
@@ -29,6 +29,8 @@ sealed class Configurator : IConfigurator {
     containerDefinition.Bind<InventoryScriptableComponent.EmptyingStatusBehavior>().AsTransient();
     containerDefinition.Bind<InventoryScriptableComponent.InventoryChangeTracker>().AsTransient();
     containerDefinition.Bind<ManufactoryScriptableComponent>().AsSingleton();
+    containerDefinition.Bind<NotificationsScriptableComponent>().AsSingleton();
+    containerDefinition.Bind<NotificationsScriptableComponent.StatusController>().AsTransient();
     containerDefinition.Bind<PausableScriptableComponent>().AsSingleton();
     containerDefinition.Bind<PlantableScriptableComponent>().AsSingleton();
     containerDefinition.Bind<PlantableScriptableComponent.PlantableTracker>().AsTransient();